+++ src/types/CardSet.ts
@@ -1,6 +1,0 @@
-import { Card } from '../entity/Card';
-
-export type CardSet = {
-    name: string;
-    cards: Card[];
-}