+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Version 0.4.0
+- Add endpoint for getting a card's image
+- Add endpoint for getting multiple cards by name at once
+
 ## Version 0.3.0
 - Add all card data and resources (images, csv sources, csv & sql exports)
 - Add 'cards' endpoint that returns all cards JSON